+++ src/app.ts
@@ -1,5 +1,5 @@
 import { Engine } from './engine';
-import { MapTile } from './gobj/maptile';
+import { FloorMap, MapTile } from './gobj/map';
 import { Player } from './gobj/player';
 import { Color } from './lib/color';
 import { Size } from './lib/size';
@@ -9,8 +9,8 @@ const game = new Engine(new Size(160, 144), new Size(800, 600), 60, new Color(0,
 
 window.addEventListener('load', () => {
     const player = new Player(new Vec2(0, 0));
-    const mapTile = new MapTile('water', new Vec2(0, 0)),
-        mapTile2 = new MapTile('grass', new Vec2(16, 0));
+    let map = new FloorMap([]);
+    fetch('data/map-test.json').then(value => value.json()).then(json => map = new FloorMap(json));
     game.start(document.body, f => {
         if (game.isKeyDown('a')) {
             player.walk('left');
@@ -23,6 +23,6 @@ window.addEventListener('load', () => {
         } else {
             player.walk('none');
         }
-        return [player, mapTile, mapTile2];
+        return [player, map];
     }, console.log);
 });